Fix Mac.

+++ arangod/RocksDBEngine/RocksDBMetaCollection.cpp
@@ -578,8 +578,8 @@ void RocksDBMetaCollection::removeRevisionTreeBlocker(TRI_voc_tid_t transactionI
 }
 
 void RocksDBMetaCollection::bufferUpdates(rocksdb::SequenceNumber seq,
-                                          std::vector<std::size_t>&& inserts,
-                                          std::vector<std::size_t>&& removals) {
+                                          std::vector<std::uint64_t>&& inserts,
+                                          std::vector<std::uint64_t>&& removals) {
   if (!_logicalCollection.syncByRevision()) {
     return;
   }
